requirements.txt
opencv_python==4.6.0.66
pyfiglet==0.8.post1
pypotrace==0.3
inquirer==-2.9.2
